+++ burn-wgpu/src/codegen/kernel.rs
@@ -1,272 +1,8 @@
-use super::dialect::gpu::Scope;
-use crate::codegen::dialect::gpu::{
-    Binding, ComputeShader, Elem, Item, Location, Variable, Vectorization, Visibility,
-    WorkgroupSize,
-};
 use crate::compute::StaticKernel;
 use crate::element::JitElement;
-use crate::fusion::tracing::CompilingInfo;
 use crate::kernel::{elemwise_workgroup, StaticKernelSource, WORKGROUP_DEFAULT};
 use crate::Runtime;
 use burn_compute::client::ComputeClient;
-use std::marker::PhantomData;
-
-/// Kernel creation input phase, see [kernel codegen](ElemWiseKernelCodegen) for more details.
-pub struct InputPhase;
-/// Kernel creation output phase, see [kernel codegen](ElemWiseKernelCodegen) for more details.
-pub struct OutputPhase;
-/// Kernel compilation phase, see [kernel codegen](ElemWiseKernelCodegen) for more details.
-pub struct CompilationPhase;
-
-#[derive(new, Clone, Copy)]
-pub struct InplaceMapping {
-    pub position_input: usize,
-    pub position_output: usize,
-}
-
-/// Allows to create custom wgsl kernels based on configured inputs, body and outputs.
-///
-/// This type has 4 phases that must be executed in order, but no worry the type system won't allow
-/// you to make mistakes.
-///
-///   1. [Input Phase](InputPhase)
-///     This phase focuses on registering the input arrays and scalars that are going to be used by
-///     the kernel.
-///   2. [Body Phase](BodyPhase)
-///     After the input phase is done, all the operations that happen in the body must be
-///     registered.
-///   3. [Output Phase](OutputPhase)
-///     This step focuses on registering all output arrays or inputs that the kernel needs to write to.
-///   4. [Compilation Phase](CompilationPhase)
-///     Now that all other phases are completed, we can actually compile the kernel.
-pub struct ElemWiseKernelCodegen<Phase = InputPhase> {
-    scope: Scope,
-    input_bindings: Vec<Binding>,
-    output_bindings: Vec<Binding>,
-    named_bindings: Vec<(String, Binding)>,
-    vectorization: Vectorization,
-    mappings_inplace: Vec<InplaceMapping>,
-    workgroup_size: WorkgroupSize,
-    _phase: PhantomData<Phase>,
-}
-
-#[derive(Clone)]
-pub enum Input {
-    Array {
-        item: Item,
-        visibility: Visibility,
-        strategy: ReadingStrategy,
-    },
-    Scalar {
-        elem: Elem,
-        size: usize,
-    },
-}
-
-#[derive(Clone)]
-pub enum ReadingStrategy {
-    /// Each element will be read in a way to be compatible with the output layout.
-    OutputLayout,
-    /// Keep the current layout.
-    Plain,
-}
-
-#[derive(Clone)]
-pub enum Output {
-    Array { item: Item, local: u16 },
-    Input { item: Item, input: u16, local: u16 },
-}
-
-impl ElemWiseKernelCodegen<InputPhase> {
-    pub fn new(scope: Scope) -> Self {
-        Self {
-            scope,
-            input_bindings: Default::default(),
-            output_bindings: Default::default(),
-            named_bindings: Default::default(),
-            vectorization: Vectorization::Scalar,
-            mappings_inplace: Default::default(),
-            workgroup_size: Default::default(),
-            _phase: PhantomData,
-        }
-    }
-
-    #[allow(dead_code)]
-    pub fn vectorize(mut self, vectorization: Vectorization) -> Self {
-        self.vectorization = vectorization;
-        self.scope.vectorize(self.vectorization);
-        self
-    }
-
-    #[allow(dead_code)]
-    pub fn inplace(mut self, mappings: &[InplaceMapping]) -> Self {
-        self.mappings_inplace = mappings.to_vec();
-        self
-    }
-
-    /// Register the inputs used by the kernel.
-    pub fn inputs(mut self, inputs: &[Input]) -> ElemWiseKernelCodegen<OutputPhase> {
-        for input in inputs {
-            match input {
-                Input::Array {
-                    item,
-                    visibility,
-                    strategy: _,
-                } => {
-                    let item = item.vectorize(self.vectorization);
-
-                    self.input_bindings.push(Binding {
-                        item: bool_item(item),
-                        visibility: *visibility,
-                        location: Location::Storage,
-                        size: None,
-                    });
-                }
-                Input::Scalar { elem, size } => {
-                    let elem = bool_elem(*elem);
-
-                    self.named_bindings.push((
-                        format!("scalars_{}", elem),
-                        Binding {
-                            item: Item::Scalar(elem),
-                            visibility: Visibility::Read,
-                            location: Location::Storage,
-                            size: Some(*size),
-                        },
-                    ));
-                }
-            }
-        }
-
-        ElemWiseKernelCodegen {
-            scope: self.scope,
-            input_bindings: self.input_bindings,
-            output_bindings: self.output_bindings,
-            named_bindings: self.named_bindings,
-            vectorization: self.vectorization,
-            mappings_inplace: self.mappings_inplace,
-            workgroup_size: self.workgroup_size,
-            _phase: PhantomData,
-        }
-    }
-}
-
-impl ElemWiseKernelCodegen<OutputPhase> {
-    /// Register the outputs with their local variable index.
-    ///
-    /// Note that the index corresponds to the registered [operator](Operator) number at the
-    /// [body phase](BodyPhase).
-    /// So the 4th operator registered creates the local variable 3 (N-1, since the 1th index is 0).
-    pub fn outputs(mut self, outputs: &[Output]) -> ElemWiseKernelCodegen<CompilationPhase> {
-        let mut index = 0;
-
-        let mut outputs = outputs.to_vec();
-
-        for mapping in self.mappings_inplace.iter() {
-            match outputs.get_mut(mapping.position_output) {
-                Some(output) => match output {
-                    Output::Array { item, local } => {
-                        *output = Output::Input {
-                            item: *item,
-                            input: mapping.position_input as u16,
-                            local: *local,
-                        };
-                    }
-                    Output::Input {
-                        item: _,
-                        input: _,
-                        local: _,
-                    } => continue,
-                },
-                None => continue,
-            }
-
-            if let Some(binding) = self.input_bindings.get_mut(mapping.position_input) {
-                binding.visibility = Visibility::ReadWrite
-            }
-        }
-
-        for array in &outputs {
-            match array {
-                Output::Array { item, local } => {
-                    let item = item.vectorize(self.vectorization);
-                    let elem_adapted = bool_item(item);
-
-                    self.output_bindings.push(Binding {
-                        item: elem_adapted,
-                        visibility: Visibility::ReadWrite,
-                        location: Location::Storage,
-                        size: None,
-                    });
-                    self.scope.write_global(
-                        Variable::Local(*local, item, self.scope.depth),
-                        Variable::Output(index, elem_adapted),
-                    );
-                    index += 1;
-                }
-                Output::Input { item, input, local } => {
-                    let item = item.vectorize(self.vectorization);
-
-                    self.scope.write_global(
-                        Variable::Local(*local, item, self.scope.depth),
-                        Variable::Input(*input, bool_item(item)),
-                    );
-                }
-            }
-        }
-
-        ElemWiseKernelCodegen {
-            scope: self.scope,
-            input_bindings: self.input_bindings,
-            output_bindings: self.output_bindings,
-            named_bindings: self.named_bindings,
-            vectorization: self.vectorization,
-            mappings_inplace: self.mappings_inplace,
-            workgroup_size: self.workgroup_size,
-            _phase: PhantomData,
-        }
-    }
-}
-
-impl ElemWiseKernelCodegen<CompilationPhase> {
-    #[allow(dead_code)] // Only used for fusion for now.
-    pub fn workgroup_size(mut self, workgroup_size: WorkgroupSize) -> Self {
-        self.workgroup_size = workgroup_size;
-        self
-    }
-
-    /// Compile the kernel into a [compute shader](ComputeShader).
-    pub fn compile(self) -> ComputeShader {
-        let inputs = self.input_bindings;
-        let outputs = self.output_bindings;
-        let mut named = Vec::with_capacity(2);
-
-        named.push((
-            "info".to_string(),
-            Binding {
-                item: Item::Scalar(Elem::UInt),
-                visibility: Visibility::Read,
-                location: Location::Storage,
-                size: None, // We avoid putting the length here since it will force a new kernel
-                            // for each tensor rank.
-            },
-        ));
-
-        for (name, binding) in self.named_bindings.into_iter() {
-            named.push((name, binding));
-        }
-
-        ComputeShader {
-            inputs,
-            outputs,
-            named,
-            workgroup_size: self.workgroup_size,
-            body: self.scope,
-            num_workgroups: true,
-            global_invocation_id: true,
-        }
-    }
-}
 
 #[derive(new)]
 pub struct StaticHandle<'a, R: Runtime> {
@@ -361,21 +97,4 @@ pub(crate) fn calculate_num_elems_dyn_rank(shape: &[usize]) -> usize {
         num_elems *= i;
     }
     num_elems
-}
-
-fn bool_item(ty: Item) -> Item {
-    match ty {
-        Item::Vec4(elem) => Item::Vec4(bool_elem(elem)),
-        Item::Vec3(elem) => Item::Vec3(bool_elem(elem)),
-        Item::Vec2(elem) => Item::Vec2(bool_elem(elem)),
-        Item::Scalar(elem) => Item::Scalar(bool_elem(elem)),
-    }
-}
-
-fn bool_elem(elem: Elem) -> Elem {
-    match elem {
-        // U32 are used for bool tensors
-        Elem::Bool => Elem::UInt,
-        _ => elem,
-    }
 }
