+++ burn-wgpu/src/codegen/mod.rs
@@ -1,7 +1,9 @@
+mod compilation;
 pub(crate) mod compiler;
 pub(crate) mod dialect;
 
 mod kernel;
 
+pub(crate) use compilation::*;
 pub(crate) use compiler::*;
 pub(crate) use kernel::*;
